Wordpress: Update Image (#183)
* Wordpress: Update Image

The "latest" tag for wordpress is no longer based on alpine, passing through the UID and GID of the user only works with alpine. This was causing wordpress to fail to start, updated image to this alpine specific version, unfortunately unable to find a way to use "latest" but continue specifying which user to run as.

* Wordpress: Update Image, Remove User Option

Updated image back to "latest", running image as alpine specific with the user specified initially looked like it worked but did not.
Removed user option.