Adding automatic configuration of download directories for DelugeVPN, following ruTorrent procedure.
This will make delugevpn default to /downloads/torrents/deluge instead of /data/downloads, which causes *arrs to not recognise the download directory being provided.

With this change DelugeVPN should work out of the box with the *arrs. However the Labels plugin still needs to be enabled by the user in order to use *arr labels. Attempted to add this in myself but its beyond my skill level with Ansible at present.