Modify postgres role to be more customizable
Currently, you can't open the port for postgres to allow ansible to
interact and run queries from outside the container. This will open the
port to localhost only and allow you to directly communicate with the
postgres intance. This is needed for another PR, so we can run a query
against the instance.

Also, added username and password variables, so this can be added into
the accounts.yml